mango_yolo_project/requirements.txt
ultralytics>=8.0.0
scikit-learn
pyyaml
opencv-python
torch>=1.7.0
torchvision>=0.8.1
tqdm
rembg 